requirements.txt
numpy==1.19.5
pandas==1.1.5
streamlit==1.9.2
Pillow==8.4.0
altair==4.1.0
